+++ packages/components/psammead-navigation/src/DropdownNavigation/index.jsx
@@ -10,12 +10,15 @@ import {
   GEL_SPACING_DBL,
 } from '@bbc/gel-foundations/spacings';
 import Helmet from 'react-helmet';
-import { GEL_GROUP_3_SCREEN_WIDTH_MIN } from '@bbc/gel-foundations/breakpoints';
+import {
+  GEL_GROUP_3_SCREEN_WIDTH_MIN,
+  GEL_GROUP_B_MIN_WIDTH,
+} from '@bbc/gel-foundations/breakpoints';
 import { getPica } from '@bbc/gel-foundations/typography';
 import { scriptPropType } from '@bbc/gel-foundations/prop-types';
 import { getSansRegular } from '@bbc/psammead-styles/font-styles';
 
-const MENU_ICON_SIDE_LENGTH = '2.75rem'; // 44px
+export const NAV_BAR_TOP_BOTTOM_SPACING = 0.75; // 12px
 
 const getStyles = dir => {
   const direction = dir === 'ltr' ? 'left' : 'right';
@@ -115,15 +118,37 @@ const iconBorder = css`
   border: ${GEL_SPACING_HLF} solid ${C_WHITE};
 `;
 
-const MenuButton = styled.button`
-  width: ${MENU_ICON_SIDE_LENGTH};
-  height: ${MENU_ICON_SIDE_LENGTH};
+// The sideLength of the button should be
+//  line height + top padding + bottom padding
+const calculateButtonSide = lineHeight =>
+  lineHeight / 16 + NAV_BAR_TOP_BOTTOM_SPACING * 2;
+
+const getButtonDimensions = lineHeight =>
+  `height: ${calculateButtonSide(lineHeight)}rem;
+  width: ${calculateButtonSide(lineHeight)}rem;`;
+
+// React allows us to put non-standard HTML attributes onto elements
+// whereas styled-components filters them out.
+// The `on` amp attribute is classed as non-standard, therefore we
+// make a react element which spreads the props as attributes
+// and style the React button.
+// For some reason the styled-components filtering only happens on SSR
+// so is not visible in tests or storybook, only once everything is
+// integrated into simorgh.
+// eslint-disable-next-line react/prop-types
+const Button = ({ script, ...props }) => <button type="button" {...props} />;
+
+const MenuButton = styled(Button)`
   position: relative;
   padding: 0;
   margin: 0;
-  border: 0;
   background-color: transparent;
+  border: 0;
+
   ${({ dir }) => (dir === 'ltr' ? `float: left;` : `float: right;`)}
+  ${({ script }) =>
+    script &&
+    getButtonDimensions(script.pica.groupA.lineHeight)}
 
   &:hover,
   &:focus {
@@ -137,20 +162,29 @@ const MenuButton = styled.button`
     display: none;
     visibility: hidden;
   }
+  @media (min-width: ${GEL_GROUP_B_MIN_WIDTH}rem) {
+    ${({ script }) =>
+      script && getButtonDimensions(script.pica.groupB.lineHeight)}
+  }
+
+  & svg {
+    vertical-align: middle;
+  }
 `;
 
 export const CanonicalMenuButton = ({
   announcedText,
   isOpen,
-  onOpen,
-  onClose,
+  onClick,
   dir,
+  script,
 }) => (
   <MenuButton
     aria-label={announcedText}
-    onClick={isOpen ? onClose : onOpen}
+    onClick={onClick}
     aria-expanded={isOpen ? 'true' : 'false'}
     dir={dir}
+    script={script}
   >
     {isOpen ? navigationIcons.cross : navigationIcons.hamburger}
   </MenuButton>
@@ -158,10 +192,10 @@ export const CanonicalMenuButton = ({
 
 CanonicalMenuButton.propTypes = {
   announcedText: string.isRequired,
-  onOpen: func.isRequired,
-  onClose: func.isRequired,
+  onClick: func.isRequired,
   isOpen: bool.isRequired,
   dir: oneOf(['ltr', 'rtl']),
+  script: shape(scriptPropType).isRequired,
 };
 
 CanonicalMenuButton.defaultProps = {
@@ -178,43 +212,45 @@ const AmpHead = () => (
   </Helmet>
 );
 
-export const AmpMenuButton = ({ announcedText, onToggle, dir }) => {
-  const expandedHandler =
-    'tap:AMP.setState({ menuState: { expanded: !menuState.expanded }})';
+const expandedHandler =
+  'AMP.setState({ menuState: { expanded: !menuState.expanded }})';
 
-  return (
-    <>
-      <AmpHead />
-      <amp-state id="menuState">
-        <script type="application/json">
-          {JSON.stringify({
-            expanded: false,
-          })}
-        </script>
-      </amp-state>
-      <MenuButton
-        aria-label={announcedText}
-        aria-expanded="false"
-        data-amp-bind-aria-expanded='menuState.expanded ? "true" : "false"'
-        on={`${expandedHandler};${onToggle}`}
-        dir={dir}
-      >
-        {cloneElement(navigationIcons.hamburger, {
-          'data-amp-bind-hidden': 'menuState.expanded',
-        })}
-        {cloneElement(navigationIcons.cross, {
-          hidden: true,
-          'data-amp-bind-hidden': '!menuState.expanded',
-        })}
-      </MenuButton>
-    </>
-  );
-};
+const initialState = { expanded: false };
+
+export const AmpMenuButton = ({ announcedText, onToggle, dir, script }) => (
+  <>
+    <AmpHead />
+    <amp-state id="menuState">
+      <script
+        type="application/json"
+        /* eslint-disable-next-line react/no-danger */
+        dangerouslySetInnerHTML={{ __html: JSON.stringify(initialState) }}
+      />
+    </amp-state>
+    <MenuButton
+      aria-label={announcedText}
+      aria-expanded="false"
+      data-amp-bind-aria-expanded='menuState.expanded ? "true" : "false"'
+      on={`tap:${expandedHandler},${onToggle}`}
+      dir={dir}
+      script={script}
+    >
+      {cloneElement(navigationIcons.hamburger, {
+        'data-amp-bind-hidden': 'menuState.expanded',
+      })}
+      {cloneElement(navigationIcons.cross, {
+        hidden: true,
+        'data-amp-bind-hidden': '!menuState.expanded',
+      })}
+    </MenuButton>
+  </>
+);
 
 AmpMenuButton.propTypes = {
   announcedText: string.isRequired,
   onToggle: string.isRequired,
   dir: oneOf(['ltr', 'rtl']),
+  script: shape(scriptPropType).isRequired,
 };
 
 AmpMenuButton.defaultProps = {