+++ packages/components/psammead-navigation/src/DropdownNavigation/index.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import { render, fireEvent, getByRole } from '@testing-library/react';
-import { latin } from '@bbc/gel-foundations/scripts';
+import { latin, arabic } from '@bbc/gel-foundations/scripts';
 import {
   Dropdown,
   DropdownUl,
@@ -35,31 +35,29 @@ const dropdownList = (
 
 describe('Canonical', () => {
   describe('Open menu button', () => {
-    it('should call onClose handler when clicked', () => {
-      const mockOnClose = jest.fn();
-      const mockOnOpen = jest.fn();
+    it('should call onClick handler when clicked', () => {
+      const mockOnClick = jest.fn();
       const { container } = render(
         <CanonicalMenuButton
           announcedText="Menu"
-          onOpen={mockOnOpen}
+          onClick={mockOnClick}
           isOpen
-          onClose={mockOnClose}
+          script={latin}
         />,
       );
       const menuButton = getByRole(container, 'button');
 
       fireEvent.click(menuButton);
-      expect(mockOnClose).toHaveBeenCalledTimes(1);
-      expect(mockOnOpen).not.toHaveBeenCalled();
+      expect(mockOnClick).toHaveBeenCalledTimes(1);
     });
 
     it('should have aria-expanded set as true', () => {
       const { container } = render(
         <CanonicalMenuButton
           announcedText="Menu"
-          onOpen={() => {}}
+          onClick={() => {}}
           isOpen
-          onClose={() => {}}
+          script={latin}
         />,
       );
       const menuButton = getByRole(container, 'button');
@@ -70,39 +68,49 @@ describe('Canonical', () => {
       'should render correctly',
       <CanonicalMenuButton
         announcedText="Menu"
-        onOpen={() => {}}
+        onClick={() => {}}
         isOpen
-        onClose={() => {}}
+        script={latin}
+        dir="ltr"
+      />,
+    );
+
+    shouldMatchSnapshot(
+      'should render rtl correctly',
+      <CanonicalMenuButton
+        announcedText="Menu"
+        onClick={() => {}}
+        isOpen
+        script={arabic}
+        dir="rtl"
       />,
     );
   });
 
   describe('Closed menu button', () => {
-    it('should call onOpen handler when clicked', () => {
-      const mockOnClose = jest.fn();
-      const mockOnOpen = jest.fn();
+    it('should call onClick handler when clicked', () => {
+      const mockOnClick = jest.fn();
       const { container } = render(
         <CanonicalMenuButton
           announcedText="Menu"
-          onOpen={mockOnOpen}
+          onClick={mockOnClick}
           isOpen={false}
-          onClose={mockOnClose}
+          script={latin}
         />,
       );
       const menuButton = getByRole(container, 'button');
 
       fireEvent.click(menuButton);
-      expect(mockOnOpen).toHaveBeenCalledTimes(1);
-      expect(mockOnClose).not.toHaveBeenCalled();
+      expect(mockOnClick).toHaveBeenCalledTimes(1);
     });
 
     it('should have aria-expanded set as false', () => {
       const { container } = render(
         <CanonicalMenuButton
           announcedText="Menu"
-          onOpen={() => {}}
+          onClick={() => {}}
           isOpen={false}
-          onClose={() => {}}
+          script={latin}
         />,
       );
       const menuButton = getByRole(container, 'button');
@@ -113,9 +121,21 @@ describe('Canonical', () => {
       'should render correctly',
       <CanonicalMenuButton
         announcedText="Menu"
-        onOpen={() => {}}
+        onClick={() => {}}
         isOpen={false}
-        onClose={() => {}}
+        script={latin}
+        dir="ltr"
+      />,
+    );
+
+    shouldMatchSnapshot(
+      'should render rtl correctly',
+      <CanonicalMenuButton
+        announcedText="Menu"
+        onClick={() => {}}
+        isOpen={false}
+        script={arabic}
+        dir="rtl"
       />,
     );
   });
@@ -131,6 +151,21 @@ describe('Dropdown navigation', () => {
 describe('AMP Menu Button', () => {
   shouldMatchSnapshot(
     'should render correctly',
-    <AmpMenuButton announcedText="Menu" onToggle="" />,
+    <AmpMenuButton
+      announcedText="Menu"
+      onToggle="other-element.toggleVisibility"
+      script={latin}
+      dir="ltr"
+    />,
+  );
+
+  shouldMatchSnapshot(
+    'should render rtl correctly',
+    <AmpMenuButton
+      announcedText="Menu"
+      onToggle="other-element.toggleVisibility"
+      script={arabic}
+      dir="rtl"
+    />,
   );
 });