+++ scripts/medea/getDependabotPRs/index.js
@@ -3,6 +3,7 @@ const axios = require('axios');
 const authToken = 'process.env.GITHUB_TOKEN';
 
 const getPullRequests = async () => {
+  console.log('Getting Dependabot Pull Requests');
   const response = await axios.get(
     'https://api.github.com/repos/bbc/psammead/pulls',
     {
@@ -13,9 +14,7 @@ const getPullRequests = async () => {
     },
   );
   const pullRequestData = response.data;
-  return pullRequestData.filter(function(one) {
-    return one != null;
-  });
+  return pullRequestData;
 };
 
 module.exports = getPullRequests;